Revert "Launcher architecture for the viewer py"
This reverts commit 9558268155520362b7945d386bfff121541e9c80.
